django_app/requirements.txt
asgiref==3.5.2
Django==3.1.14
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
